rfc6350.vcard: Fixed minor bug that raised `DeprecationWarning` in test suite
The tests were throwing the `DeprecationWarning` is because they were passing `bytes` objects to `format()`.

Given Python 3’s stricter str/bytes separation, it was raising a Warning that clearly indicated something to do with `format()`. But when I looked at that line of code, the message didn’t really make sense on first glance.

A little ghetto-style error printing, and I figured out what was going on. ☺